requirements.txt
Flask
flask-restful
mysql-connector-python
matplotlib
sklearn
pandas
sqlalchemy
pymysql